Collapse storage unit pickers to one Form component + local parent picker
Form/ now exposes a single <StorageUnit> picker (form-bound when given a name,
controlled otherwise), built on CreatableCombobox, plus the shared
useStorageUnits/useStorageUnitsTree data hooks. The hierarchical drill-down —
used only when choosing where a unit sits in the tree — moves into
StorageUnits/StorageUnitParentSelect, local to the Storage Unit form, with the
create flow and inline mode dropped (and one circular import removed).

Deletes Form/StorageUnitSelect and Form/StorageUnitDrillSelect. Restores the
value-forwarding in form mode that the drill era had dropped.
