+++ apps/erp/app/components/Form/StorageUnit.tsx
@@ -1,97 +1,63 @@
-// `<StorageUnit>` — the historical leaf/bin picker API. Delegates to the
-// `CreatableCombobox`-based `StorageUnitSelect`:
-// - with `name`     -> form-bound variant
-// - with no `name`  -> controlled variant (`storageUnitId` + `onChange`)
+// `<StorageUnit>` — the storage-unit (bin) picker. One component, two modes:
+// - with `name`    -> form-bound (`@carbon/form` CreatableCombobox)
+// - without `name` -> controlled (`value` + `onChange`) for table cells
 //
-// The hierarchical drill-down (`StorageUnitDrillSelect`) is a separate
-// component used only by the Parent picker.
+// Options are the *leaf* storage units (the storable bins), each with a helper
+// showing on-hand quantity (when `itemId` is given) or its hierarchy path.
+// Built on the canonical `CreatableCombobox`, like every other entity picker.
+//
+// Choosing where a unit sits in the tree (a non-leaf target) is a different,
+// single-purpose interaction handled by `StorageUnitParentSelect`, local to the
+// Storage Unit form.
 
-import type { ComboboxProps } from "@carbon/form";
-import { forwardRef, useEffect, useMemo } from "react";
+import { CreatableCombobox } from "@carbon/form";
+import {
+  CreatableCombobox as CreatableComboboxBase,
+  useDisclosure
+} from "@carbon/react";
+import type { MouseEventHandler } from "react";
+import { useEffect, useMemo, useRef, useState } from "react";
 import { useFetcher } from "react-router";
 import type { getStorageUnitsList } from "~/modules/inventory";
+import StorageUnitForm from "~/modules/inventory/ui/StorageUnits/StorageUnitForm";
 import type { ListItem } from "~/types";
 import { path } from "~/utils/path";
-import {
-  StorageUnitSelect,
-  StorageUnitSelectControl
-} from "./StorageUnitSelect";
 
-type StorageUnitSelectProps = Omit<
-  ComboboxProps,
-  "options" | "onChange" | "inline"
-> & {
-  locationId?: string;
-  /** Per-item qty hints shown as option helper text when provided. */
-  itemId?: string;
-  /** Controlled value (used when `name` is absent). */
-  storageUnitId?: string | null;
-  inline?: boolean;
-  onChange?: (storageUnit: ListItem | null) => void;
+// ---------------------------------------------------------------------------
+// Data hooks (shared)
+// ---------------------------------------------------------------------------
+
+export type StorageUnitTreeRow = {
+  id: string;
+  name: string;
+  parentId: string | null;
+  depth: number;
+  ancestorPath: string[];
 };
 
-const StorageUnit = forwardRef<HTMLDivElement, StorageUnitSelectProps>(
-  (props, _ref) => {
-    const {
-      name,
-      storageUnitId,
-      label,
-      inline,
-      locationId,
-      itemId,
-      isReadOnly,
-      isOptional,
-      onChange,
-      ...rest
-    } = props;
+/**
+ * Recursive storage-unit tree for a single location. One round-trip per
+ * locationId change; consumers cache by reference identity. Also used by the
+ * parent picker and storage-rule value inputs.
+ */
+export function useStorageUnitsTree(locationId?: string | null) {
+  const fetcher = useFetcher<{
+    data: StorageUnitTreeRow[] | null;
+    error: unknown;
+  }>();
 
-    // Controlled (non-form) usage: table cells / line editors that own the
-    // value and update via `onChange`.
-    if (!name) {
-      return (
-        <StorageUnitSelectControl
-          locationId={locationId}
-          value={storageUnitId}
-          itemId={itemId}
-          isReadOnly={isReadOnly}
-          placeholder={
-            typeof rest.placeholder === "string" ? rest.placeholder : undefined
-          }
-          onChange={(unit) => onChange?.(unit)}
-        />
-      );
-    }
+  // biome-ignore lint/correctness/useExhaustiveDependencies: fetcher identity changes every render
+  useEffect(() => {
+    if (locationId) fetcher.load(path.to.api.storageUnitsTree(locationId));
+  }, [locationId]);
 
-    return (
-      <StorageUnitSelect
-        name={name}
-        label={label}
-        inline={inline}
-        helperText={
-          typeof rest.helperText === "string" ? rest.helperText : undefined
-        }
-        locationId={locationId}
-        itemId={itemId}
-        isReadOnly={isReadOnly}
-        isOptional={isOptional}
-        placeholder={
-          typeof rest.placeholder === "string" ? rest.placeholder : undefined
-        }
-        onChange={(unit) => onChange?.(unit)}
-      />
-    );
-  }
-);
-StorageUnit.displayName = "StorageUnit";
+  return fetcher.data?.data ?? [];
+}
 
-export default StorageUnit;
-
-// ---------------------------------------------------------------------------
-// useStorageUnits — kept exported because table cells / non-form callsites
-// still pull the flat list (with optional per-item qty hints). The drill
-// picker has its own tree fetch; this hook stays as the flat fallback.
-// ---------------------------------------------------------------------------
-
+/**
+ * Flat storage-unit list for a location, optionally with per-item quantity
+ * helpers. Kept for non-leaf-aware callsites that pull their own options.
+ */
 export function useStorageUnits(locationId?: string, itemId?: string) {
   const storageUnitsFetcher =
     useFetcher<Awaited<ReturnType<typeof getStorageUnitsList>>>();
@@ -124,7 +90,7 @@ export function useStorageUnits(locationId?: string, itemId?: string) {
         (storageUnit: any) => !storageUnitIdsWithQuantities.has(storageUnit.id)
       );
 
-      const combinedStorageUnits = [
+      return [
         ...storageUnitsWithQuantities.map((c: any) => ({
           value: c.id,
           label: c.name,
@@ -135,8 +101,6 @@ export function useStorageUnits(locationId?: string, itemId?: string) {
           label: c.name
         }))
       ];
-
-      return combinedStorageUnits;
     }
 
     return (
@@ -156,3 +120,221 @@ export function useStorageUnits(locationId?: string, itemId?: string) {
 
   return { options, data: storageUnitsFetcher.data };
 }
+
+type StorageUnitOption = { value: string; label: string; helper?: string };
+
+/**
+ * Per-unit on-hand quantities for `itemId`, keyed by storage-unit id. Only
+ * fetches when both ids are present, so item-less pickers don't pay for it.
+ */
+function useStorageUnitQuantities(locationId?: string | null, itemId?: string) {
+  const fetcher = useFetcher<{ data: { id: string; quantity: number }[] }>();
+
+  // biome-ignore lint/correctness/useExhaustiveDependencies: fetcher identity changes every render
+  useEffect(() => {
+    if (locationId && itemId) {
+      fetcher.load(path.to.api.storageUnitsWithQuantities(locationId, itemId));
+    }
+  }, [locationId, itemId]);
+
+  return useMemo(() => {
+    const m = new Map<string, number>();
+    for (const s of fetcher.data?.data ?? []) m.set(s.id, s.quantity);
+    return m;
+  }, [fetcher.data]);
+}
+
+/**
+ * Options for the leaf bins in a location (nodes with no children — the
+ * storable locations). `helper` shows the on-hand quantity when `itemId` is
+ * given, otherwise the hierarchy path to disambiguate same-named bins.
+ */
+export function useStorageUnitLeafOptions(
+  locationId?: string | null,
+  itemId?: string
+): StorageUnitOption[] {
+  const rows = useStorageUnitsTree(locationId);
+  const quantities = useStorageUnitQuantities(locationId, itemId);
+
+  return useMemo(() => {
+    const byId = new Map(rows.map((r) => [r.id, r]));
+    const parentIds = new Set(
+      rows.map((r) => r.parentId).filter((id): id is string => Boolean(id))
+    );
+
+    return rows
+      .filter((r) => !parentIds.has(r.id))
+      .map((r) => {
+        const ancestorPath = (r.ancestorPath ?? [])
+          .slice(0, -1)
+          .map((id) => byId.get(id)?.name)
+          .filter(Boolean)
+          .join(" / ");
+        return {
+          value: r.id,
+          label: r.name,
+          helper: itemId
+            ? `Qty: ${quantities.get(r.id) ?? 0}`
+            : ancestorPath || undefined
+        };
+      })
+      .sort((a, b) => a.label.localeCompare(b.label));
+  }, [rows, quantities, itemId]);
+}
+
+// ---------------------------------------------------------------------------
+// Create-new-unit flow
+// ---------------------------------------------------------------------------
+
+/**
+ * Shared "+ New storage unit" flow. `onCreateOption` seeds the form with the
+ * typed text; closing the modal re-clicks the trigger so the combobox reopens
+ * with the freshly created unit selectable.
+ */
+function useNewStorageUnitModal(locationId?: string | null) {
+  const modal = useDisclosure();
+  const [created, setCreated] = useState("");
+  const triggerRef = useRef<HTMLButtonElement>(null);
+
+  const node =
+    modal.isOpen && locationId ? (
+      <StorageUnitForm
+        type="modal"
+        locationId={locationId}
+        onClose={() => {
+          setCreated("");
+          modal.onClose();
+          triggerRef.current?.click();
+        }}
+        initialValues={{ name: created, locationId, storageTypeIds: [] }}
+      />
+    ) : null;
+
+  const onCreateOption = (value: string) => {
+    setCreated(value);
+    modal.onOpen();
+  };
+
+  return { triggerRef, onCreateOption, node };
+}
+
+const storageUnitPreview = (
+  value: string,
+  options: { value: string; label: string | JSX.Element }[]
+) => {
+  const match = options.find((o) => o.value === value);
+  return <span className="text-sm">{match?.label ?? ""}</span>;
+};
+
+const labelToString = (label: string | JSX.Element) =>
+  typeof label === "string" ? label : "";
+
+// ---------------------------------------------------------------------------
+// Component
+// ---------------------------------------------------------------------------
+
+type StorageUnitProps = {
+  /** Provide `name` for form-bound usage; omit it for controlled usage. */
+  name?: string;
+  /** Form-mode seed value, or the controlled value when `name` is absent. */
+  value?: string | null;
+  label?: string;
+  helperText?: string;
+  placeholder?: string;
+  isReadOnly?: boolean;
+  isOptional?: boolean;
+  disabled?: boolean;
+  className?: string;
+  onClick?: MouseEventHandler<HTMLButtonElement>;
+  locationId?: string | null;
+  itemId?: string;
+  /** Render a compact inline preview (table cells) once a value is set. */
+  inline?: boolean;
+  /** Show the "+ New storage unit" create option. Defaults to `true`. */
+  allowCreate?: boolean;
+  onChange?: (storageUnit: ListItem | null) => void;
+};
+
+const toListItem = (id: string, options: StorageUnitOption[]): ListItem => ({
+  id,
+  name: labelToString(options.find((o) => o.value === id)?.label ?? "")
+});
+
+function StorageUnit({
+  name,
+  value,
+  label,
+  helperText,
+  placeholder = "Select storage unit",
+  isReadOnly,
+  isOptional,
+  disabled,
+  className,
+  onClick,
+  locationId,
+  itemId,
+  inline,
+  allowCreate = true,
+  onChange
+}: StorageUnitProps) {
+  const options = useStorageUnitLeafOptions(locationId, itemId);
+  const { triggerRef, onCreateOption, node } =
+    useNewStorageUnitModal(locationId);
+  const readOnly = isReadOnly || disabled;
+
+  if (name) {
+    return (
+      <>
+        <CreatableCombobox
+          ref={triggerRef}
+          name={name}
+          value={value ?? undefined}
+          options={options}
+          label={label ?? "Storage Unit"}
+          helperText={helperText}
+          placeholder={placeholder}
+          isReadOnly={readOnly}
+          isOptional={isOptional}
+          className={className}
+          onClick={onClick}
+          inline={inline ? storageUnitPreview : undefined}
+          onCreateOption={allowCreate ? onCreateOption : undefined}
+          onChange={(option) =>
+            onChange?.(
+              option
+                ? { id: option.value, name: labelToString(option.label) }
+                : null
+            )
+          }
+        />
+        {node}
+      </>
+    );
+  }
+
+  if (!locationId) return null;
+
+  return (
+    <>
+      <CreatableComboboxBase
+        ref={triggerRef}
+        options={options}
+        value={value ?? undefined}
+        isReadOnly={readOnly}
+        isClearable
+        placeholder={placeholder}
+        className={className}
+        onClick={onClick}
+        onCreateOption={allowCreate ? onCreateOption : undefined}
+        onChange={(selected) =>
+          onChange?.(selected ? toListItem(selected, options) : null)
+        }
+      />
+      {node}
+    </>
+  );
+}
+
+StorageUnit.displayName = "StorageUnit";
+
+export default StorageUnit;
