+++ apps/erp/app/modules/inventory/ui/StorageUnits/StorageUnitForm.tsx
@@ -22,10 +22,10 @@ import {
   StorageTypes,
   Submit
 } from "~/components/Form";
-import { StorageUnitDrillSelectField } from "~/components/Form/StorageUnitDrillSelect";
 import { usePermissions } from "~/hooks";
 import { storageUnitValidator } from "~/modules/inventory";
 import { path } from "~/utils/path";
+import { StorageUnitParentSelect } from "./StorageUnitParentSelect";
 
 type StorageUnitFormProps = {
   locationId: string;
@@ -93,7 +93,7 @@ const StorageUnitForm = ({
                   name="locationId"
                   label={t`Location`}
                 />
-                <StorageUnitDrillSelectField
+                <StorageUnitParentSelect
                   name="parentId"
                   label={t`Parent Storage Unit`}
                   locationId={locationId}
