+++ apps/erp/app/modules/items/ui/Item/PickMethodForm.tsx
@@ -44,7 +44,7 @@ import {
   ShelfLifeStartTiming,
   Submit
 } from "~/components/Form";
-import { StorageUnitSelect } from "~/components/Form/StorageUnitSelect";
+import StorageUnit from "~/components/Form/StorageUnit";
 import { usePermissions, useSettings, useUser } from "~/hooks";
 import type { ListItem } from "~/types";
 import { path } from "~/utils/path";
@@ -181,7 +181,7 @@ const PickMethodForm = ({
           <Hidden name="itemId" />
           <Hidden name="locationId" />
           <div className="grid grid-cols-1 lg:grid-cols-3 gap-x-8 gap-y-4 w-full">
-            <StorageUnitSelect
+            <StorageUnit
               name="defaultStorageUnitId"
               label={t`Default Storage Unit`}
               locationId={initialValues.locationId}
